Remove CATEGORY keyword from FLAG_CATEGORY column
Remove the keyword from the column description.
This empty Vector<String> gets converted to a python dictionary as
'FLAG_CATEGORY' : {
     ...
     keywords': {'CATEGORY' : []},
     ...
}

Due to the missing type information this gets converted
into something like Vector<int> when passed back to the C++
layer, which results in Table Conformance errors.
This is a reasonable solution since the C++ layer always adds
this keyword if it is missing from the MS (see addCat())