Restore macOS support
Force clang to use the C++11 standard (mandatory since casacore 3.0.0).

Look for the multithreaded version of Boost.Python first, which has a
different -mt extension on macOS. Linux does not use this extension, so
the extra checks should be safe on that platform. This avoids linking
to the non-mt version of the library (Homebrew installs both so it is
necessary to pick the right one).