Improve find_boost and find_casacore in setup
The library dirs and include dirs found in setup.py are now passed
to the compiler. Find_boost has been improved to always prefer
library names with versions over those without. So 'boost_python36'
is now preferred over 'boost_python-mt' (it was the other way around).

Including the header dir will fix issues with 'missing symbols' since
it could happen that a header from a previous boost installation was
available in a system path, which was slightly incompatible with the
boost library used at runtime.