Remove dependency on external argparse package
There are two problems with having an install_requires dependency on the
external argparse package: first, the argparse module has been included
in the Python standard library since 2.7 and 3.2, and therefore there's
no need to install it externally; and secondly, the code in
python-casacore doesn't use this dependency. In fact the only place
where argparse is used is in setup.py itself, which is a testimony that
the module shipped with the standard library is enough.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>