Explicitly add Homebrew libs to support Apple M1
In the grand old days, Homebrew installed to `/usr/local` and all was well.
Now, however, while Intel Homebrew still does this, Apple Silicon / M1 / M2
Homebrew has gone and picked a new prefix: `/opt/homebrew`. A good explanation
of the rationale is at https://earthly.dev/blog/homebrew-on-m1/. Basically,
Homebrew had to admit that Fink and MacPorts were right. :-P

The solution is to call `brew --prefix` and add those libraries explicitly
to the search list. Keep going if brew is not installed.

This addresses #252.